code/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
peft>=0.4.0
datasets>=2.12.0
scikit-learn>=1.2.0
pyyaml>=6.0
pillow>=9.0.0
numpy>=1.24.0
